Added neon to project

+++ Color_Space_Conversion/LESSON_103_CSC/C/CSC_RGB_to_YCC_01.c
@@ -280,13 +280,13 @@ static void CSC_RGB_to_YCC_neon( int row, int col) {
     scalar_vector_C3 = vdupq_n_u32(C33);
     BB_scaled = vmulq_u32(BB, scalar_vector_C3);
 
-    uint32x4_t CrCr = vaddq_u32(128 << (K), RR_scaled);
+    uint32x4_t CrCr = vaddq_u32((uint32_t)(128 << (K)), RR_scaled);
 
     CrCr = vsubq_u32(CrCr, GG_scaled);
 
     CrCr = vsubq_u32(CrCr, BB_scaled);
 
-    CrCr = vaddq_u32(CrCr, 1 << (K-1)); // rounding
+    CrCr = vaddq_u32(CrCr, (uint32_t)(1 << (K-1))); // rounding
     CrCr = vshrq_n_u32(CrCr, K);
 
     uint32_t CrCr_result[4];
